Loading data

+++ TMDB Movies/Coredata/Genre+CoreDataProperties.swift
@@ -17,7 +17,7 @@ extension Genre {
         return NSFetchRequest<Genre>(entityName: "Genre")
     }
 
-    @NSManaged public var id: Int16
+    @NSManaged public var id: Int64
     @NSManaged public var name: String?
 
 }
